+++ business/src/main/java/com/kevinguanchedarias/owgejava/business/UnitBo.java
@@ -1,17 +1,8 @@
 package com.kevinguanchedarias.owgejava.business;
 
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-import org.apache.commons.lang3.ObjectUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-
 import com.kevinguanchedarias.owgejava.dto.InterceptableSpeedGroupDto;
 import com.kevinguanchedarias.owgejava.dto.UnitDto;
+import com.kevinguanchedarias.owgejava.entity.CriticalAttack;
 import com.kevinguanchedarias.owgejava.entity.InterceptableSpeedGroup;
 import com.kevinguanchedarias.owgejava.entity.Unit;
 import com.kevinguanchedarias.owgejava.entity.UserStorage;
@@ -20,31 +11,36 @@ import com.kevinguanchedarias.owgejava.exception.SgtBackendInvalidInputException
 import com.kevinguanchedarias.owgejava.pojo.ResourceRequirementsPojo;
 import com.kevinguanchedarias.owgejava.repository.InterceptableSpeedGroupRepository;
 import com.kevinguanchedarias.owgejava.repository.UnitRepository;
+import lombok.AllArgsConstructor;
+import org.apache.commons.lang3.ObjectUtils;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-@Component
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+@Service
+@AllArgsConstructor
 public class UnitBo implements WithNameBo<Integer, Unit, UnitDto>, WithUnlockableBo<Integer, Unit, UnitDto> {
 	private static final long serialVersionUID = 8956360591688432113L;
 
-	@Autowired
-	private UnitRepository unitRepository;
+	private final UnitRepository unitRepository;
 
-	@Autowired
-	private UnlockedRelationBo unlockedRelationBo;
+	private final UnlockedRelationBo unlockedRelationBo;
 
-	@Autowired
-	private ObjectRelationBo objectRelationBo;
+	private final ObjectRelationBo objectRelationBo;
 
-	@Autowired
-	private ObtainedUnitBo obtainedUnitBo;
+	private final ObtainedUnitBo obtainedUnitBo;
 
-	@Autowired
-	private ImprovementBo improvementBo;
+	private final ImprovementBo improvementBo;
 
-	@Autowired
-	private transient InterceptableSpeedGroupRepository interceptableSpeedGroupRepository;
+	private final transient InterceptableSpeedGroupRepository interceptableSpeedGroupRepository;
 
-	@Autowired
-	private SpeedImpactGroupBo speedImpactGroupBo;
+	private final SpeedImpactGroupBo speedImpactGroupBo;
+
+	private final transient CriticalAttackBo criticalAttackBo;
 
 	@Override
 	public JpaRepository<Unit, Integer> getRepository() {
@@ -127,21 +123,6 @@ public class UnitBo implements WithNameBo<Integer, Unit, UnitDto>, WithUnlockabl
 	/**
 	 * Calculates the requirements according to the count to operate!
 	 *
-	 * @param unitId
-	 * @param count
-	 * @return
-	 * @author Kevin Guanche Darias
-	 */
-	public ResourceRequirementsPojo calculateRequirements(Integer unitId, Long count) {
-		return calculateRequirements(findByIdOrDie(unitId), count);
-	}
-
-	/**
-	 * Calculates the requirements according to the count to operate!
-	 *
-	 * @param unit
-	 * @param count
-	 * @return
 	 * @throws SgtBackendInvalidInputException can't be negative
 	 * @author Kevin Guanche Darias
 	 */
@@ -150,7 +131,7 @@ public class UnitBo implements WithNameBo<Integer, Unit, UnitDto>, WithUnlockabl
 			throw new SgtBackendInvalidInputException("Input can't be negative");
 		}
 
-		ResourceRequirementsPojo retVal = new ResourceRequirementsPojo();
+		var retVal = new ResourceRequirementsPojo();
 		retVal.setRequiredPrimary((double) (unit.getPrimaryResource() * count));
 		retVal.setRequiredSecondary((double) (unit.getSecondaryResource() * count));
 		retVal.setRequiredTime((double) (unit.getTime() * count));
@@ -165,8 +146,6 @@ public class UnitBo implements WithNameBo<Integer, Unit, UnitDto>, WithUnlockabl
 	/**
 	 * Checks if the unique unit has been build by the user
 	 *
-	 * @param user
-	 * @param unit
 	 * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
 	 */
 	public void checkIsUniqueBuilt(UserStorage user, Unit unit) {
@@ -179,21 +158,25 @@ public class UnitBo implements WithNameBo<Integer, Unit, UnitDto>, WithUnlockabl
 
 	/**
 	 *
-	 * @param interceptableSpeedGroupDtos
 	 * @since 0.10.0
 	 * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
 	 */
 	@Transactional
 	public void saveSpeedImpactGroupInterceptors(int unitId,
 			List<InterceptableSpeedGroupDto> interceptableSpeedGroupDtos) {
-		Unit unit = getOne(unitId);
+		var unit = getOne(unitId);
 		interceptableSpeedGroupRepository.deleteByUnit(unit);
 		interceptableSpeedGroupDtos.forEach(current -> {
-			InterceptableSpeedGroup interceptableSpeedGroup = new InterceptableSpeedGroup();
+			var interceptableSpeedGroup = new InterceptableSpeedGroup();
 			interceptableSpeedGroup.setUnit(unit);
 			interceptableSpeedGroup
 					.setSpeedImpactGroup(speedImpactGroupBo.getOne(current.getSpeedImpactGroup().getId()));
 			interceptableSpeedGroupRepository.save(interceptableSpeedGroup);
 		});
+	}
+
+	public CriticalAttack findUsedCriticalAttack(int unitId) {
+		var unit = findByIdOrDie(unitId);
+		return ObjectUtils.firstNonNull(unit.getCriticalAttack(), criticalAttackBo.findUsedCriticalAttack(unit.getType()));
 	}
 }