+++ game-rest/src/main/java/com/kevinguanchedarias/owgejava/rest/game/UnitRestService.java
@@ -1,25 +1,12 @@
 package com.kevinguanchedarias.owgejava.rest.game;
 
-import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-import org.hibernate.Hibernate;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.context.annotation.ApplicationScope;
-
 import com.kevinguanchedarias.owgejava.builder.SyncHandlerBuilder;
+import com.kevinguanchedarias.owgejava.business.CriticalAttackBo;
 import com.kevinguanchedarias.owgejava.business.FactionBo;
 import com.kevinguanchedarias.owgejava.business.MissionBo;
 import com.kevinguanchedarias.owgejava.business.ObtainedUnitBo;
 import com.kevinguanchedarias.owgejava.business.RequirementBo;
+import com.kevinguanchedarias.owgejava.business.UnitBo;
 import com.kevinguanchedarias.owgejava.business.UnlockedRelationBo;
 import com.kevinguanchedarias.owgejava.business.UserStorageBo;
 import com.kevinguanchedarias.owgejava.dto.ObtainedUnitDto;
@@ -32,34 +19,45 @@ import com.kevinguanchedarias.owgejava.enumerations.RequirementTargetObject;
 import com.kevinguanchedarias.owgejava.interfaces.SyncSource;
 import com.kevinguanchedarias.owgejava.pojo.DeprecationRestResponse;
 import com.kevinguanchedarias.owgejava.pojo.UnitWithRequirementInformation;
+import com.kevinguanchedarias.owgejava.responses.CriticalAttackInformationResponse;
+import lombok.AllArgsConstructor;
+import org.hibernate.Hibernate;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.context.annotation.ApplicationScope;
+
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("game/unit")
 @ApplicationScope
+@AllArgsConstructor
 public class UnitRestService implements SyncSource {
 
-	@Autowired
-	private UserStorageBo userStorageBo;
+	private final UserStorageBo userStorageBo;
 
-	@Autowired
-	private UnlockedRelationBo unlockedRelationBo;
+	private final UnlockedRelationBo unlockedRelationBo;
 
-	@Autowired
-	private MissionBo missionBo;
+	private final MissionBo missionBo;
 
-	@Autowired
-	private ObtainedUnitBo obtainedUnitBo;
+	private final ObtainedUnitBo obtainedUnitBo;
 
-	@Autowired
-	private RequirementBo requirementBo;
-
-	@Autowired
-	private FactionBo factionBo;
+	private final RequirementBo requirementBo;
+	private final FactionBo factionBo;
+	private final UnitBo unitBo;
+	private final CriticalAttackBo criticalAttackBo;
 
 	/**
-	 *
-	 * @deprecated Find in all planets instead
 	 * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
+	 * @deprecated Find in all planets instead
 	 */
 	@Deprecated(since = "0.9.0")
 	@GetMapping("findRunning")
@@ -95,6 +93,14 @@ public class UnitRestService implements SyncSource {
 		obtainedUnitDto.setUserId(userStorageBo.findLoggedIn().getId());
 		obtainedUnitBo.saveWithSubtraction(obtainedUnitDto, true);
 		return "\"OK\"";
+	}
+
+	@GetMapping("{unitId}/criticalAttack")
+	public List<CriticalAttackInformationResponse> findCriticalAttackInformation(@PathVariable int unitId) {
+		var criticalAttack = unitBo.findUsedCriticalAttack(unitId);
+		return criticalAttack == null
+				? List.of()
+				: criticalAttackBo.buildFullInformation(criticalAttack);
 	}
 
 	@Override